Individual project for university
My goal in this project was to develop a ML model which will help reduce the strain on hospital’s ICU beds where the objective is to develop a ML model to predict if a patient of confirmed COVID-19 case would require admission to the ICU.  With that objective in mind, the team has collected a decent amount of clinical data from patients i.e. the features of COVID patients, and the target (those been admitted to ICU).

The dataset is released on Kaggle platform with full data description at the following URL (https://www.kaggle.com/S%C3%ADrio-Libanes/covid19) by the team seeking interesting solutions and findings from the public.